\name{CRSPutils-package}
\alias{CRSPutils-package}
\alias{CRSPutils}
\docType{package}
\title{
\packageTitle{CRSPutils}
}
\description{
\packageDescription{CRSPutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CRSPutils}
\packageIndices{CRSPutils}
}
\author{
\packageAuthor{CRSPutils}

Maintainer: \packageMaintainer{CRSPutils}
}
%% \references{
%% ~~ Literature or other references for background information ~~
%% }
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%% ~~ the R documentation directory ~~
\keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
